# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/candidate:
    get:
      tags:
        - Candidate
      summary: 'Получить информацию об авторизованном студенте'
      operationId: 2480a493492a84a70c52416430a1d362
      responses:
        '200':
          description: 'Информация о студенте'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
    put:
      tags:
        - Candidate
      summary: 'Обновить номер телефона, о себе, скиллы студента @TODO BODY'
      operationId: 7992d4e4a9e0c02c61e4717a66494a25
      requestBody:
        content:
          application/json:
            schema:
              properties:
                'skill_ids[]':
                  type: array
                  items: { type: integer }
                about:
                  type: string
                phone:
                  type: string
              type: object
      responses:
        '200':
          description: 'Информация о студенте Обновлена'
...
  /api/projects/filter:
    get:
      tags:
        - Project
      summary: 'Получить проекты с фильтрацией'
      operationId: 3172f11dfe97cf7f45f02c67fdfc86e5
      parameters:
        -
          name: page
          in: query
          description: 'Номер страницы (отчет с нуля)'
          required: false
          schema:
            type: integer
        -
          name: skills
          in: query
          description: 'Массив ID скилов'
          required: false
          schema:
            type: array
            items:
              type: integer
        -
          name: specialties
          in: query
          description: 'Массив ID специальностей'
          required: false
          schema:
            type: array
            items:
              type: integer
        -
          name: type
          in: query
          description: 'Массив ID типов проекта'
          required: false
          schema:
            type: array
            items:
              type: integer
        -
          name: state
          in: query
          description: 'Массив ID состояний проекта'
          required: false
          schema:
            type: array
            items:
              type: integer
        -
          name: supervisor
          in: query
          description: 'Массив ID руководителей'
          required: false
          schema:
            type: array
            items:
              type: integer
        -
          name: difficulty
          in: query
          description: 'Массив с сложностями'
          required: false
          schema:
            type: array
            items:
              type: integer
        -
          name: order
          in: query
          description: 'Порядок сортировки (возрастающий - asc, убывающий - desc)'
          required: false
          schema:
            type: string
        -
          name: sortBy
          in: query
          description: 'По какому полю сортировать'
          required: false
          schema:
            type: string
        -
          name: pageSize
          in: query
          description: 'Размер страницы. Позитивное число'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: '7 проектов на странице + навигация'
          content:
            application/json:
              schema:
                properties:
                  projectCount: { description: 'Общее количество проектов', type: integer }
                  data: { description: 'Массив проектов', type: array, items: { $ref: '#/components/schemas/Project' } }
                type: object
...
components:
  schemas:
    Candidate:
      properties:
        id:
          description: 'ID студента'
          type: integer
        fio:
          description: 'ФИО студента'
          type: string
        about:
          description: 'Информация о студенте'
          type: string
        email:
          description: Почта
          type: string
        numz:
          description: 'Номер зачетной книжки'
          type: string
        phone:
          description: Телефон
          type: string
        course:
          description: Курс
          type: integer
        training_group:
          description: 'Учебная группа'
          type: string
        canSendParticipations:
          description: 'Может ли отправлять заявки на проекты'
          type: boolean
      type: object
...
    Project:
      properties:
        id:
          description: 'ID проекта'
          type: integer
        prevProjectId:
          description: 'ID предшествующего проекта или null'
          type: integer
        title:
          description: 'Название проекта'
          type: string
        places:
          description: 'Количество мест на проекте'
          type: integer
        goal:
          description: Цель
          type: string
        description:
          description: Описание
          type: string
        difficulty:
          description: Сложность
          type: integer
        date_start:
          description: 'Дата начала (2022-09-01)'
          type: string
        date_end:
          description: 'Дата окончания (2023-05-30)'
          type: string
        requirements:
          description: Требования
          type: string
        additional_inf:
          description: 'Дополнительная информация'
          type: string
        product_result:
          description: 'Продуктовынй результат'
          type: string
        study_result:
          description: 'Учебный результат'
          type: string
        supervisors:
          description: Руководители
          type: array
          items:
            $ref: '#/components/schemas/Supervisor'
        skills:
          description: 'Требуемые навыки'
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        specialities:
          description: 'Для каких специальностей'
          type: array
          items:
            $ref: '#/components/schemas/Speciality'
        supervisorsNames:
          description: 'Имена руководителей'
          type: string
        state:
          $ref: '#/components/schemas/ProjectState'
        type:
          $ref: '#/components/schemas/ProjectType'
        created_at:
          description: 'Дата создания записи в БД'
          type: string
        updated_at:
          description: 'Дата обновления записи в БД'
          type: string
      type: object
    SkillCategory:
      properties:
        id:
          description: 'ID Категории скила'
          type: integer
        name:
          description: 'Название категории скила'
          type: string
      type: object
    Skill:
      properties:
        id:
          description: 'ID скила'
          type: integer
        name:
          description: 'Название скила'
          type: string
        skillCategory:
          $ref: '#/components/schemas/SkillCategory'
      type: object
...
    Speciality:
      properties:
        id:
          description: 'ID специальности'
          type: integer
        state:
          description: 'Название состояния'
          type: string
        institute:
          description: 'Институт к которому принадлежит специальность'
          type: object
      type: object
...
    ProjectState:
      properties:
        id:
          description: 'ID Состояния проекта'
          type: integer
        state:
          description: 'Название состояния'
          type: string
      type: object
    Supervisor:
      properties:
        id:
          description: 'ID Преподавателя'
          type: integer
        fio:
          description: ФИО
          type: string
        email:
          description: Почта
          type: string
        about:
          description: 'Информация о преподавателе'
          type: string
        position:
          description: Должность
          type: string
        created_at:
          description: 'Дата создания записи в БД'
          type: string
        updated_at:
          description: 'Дата обновления записи в БД'
          type: string
      type: object
    ProjectType:
      properties:
        id:
          description: 'ID типа проекта'
          type: integer
        type:
          description: 'Название типа проекта'
          type: string
      type: object